+++ CHANGELOG.md
@@ -43,6 +43,8 @@
   list of addons that have been changed between two specified git revisions
 - New command `odoo-helper ci check-versions-git`.
   The goal of this command is to be sure that addon version number was updated.
+- New command `odoo-helper ci ensure-icons`.
+  Ensure that all addons in specified directory have icons.
 - New option `--ual` for `odoo-helper link` command.
 - New command `odoo-helper browse` that opens running odoo instance in webbrowser
 - Experimental support of Odoo 12.0